1. 메인화면 디버깅 및 복구완료 && 측정결과 int형으로 표시 완료  - 측정 부분 디버깅 필요 2. 채팅 Layout(작성완료) 3. 채팅 1:1 완료(방개념 o) 4. 그룹채팅 레이아웃(작성완료) 5. 그룹채팅 디버깅 완료(어플을 껐다가 켜도 메세지가 보내짐) 6. FCM 완료(세부 처리 필요-프로필사진, 그룹채팅 테스트 안해봄, notification을 누르면 채팅방으로 연결 필요) 7. JSON형식을 통핸 채팅 처리 8. 채팅방(닉네임) 9. 심박,호흡보내기 최신,선택(레이아웃 수정 필요) 10. 기존대화 내용 보여주는 부분(페이징 필요) 11. 측정결과 보여주는 부분 -> 로직 수정 / 쿼리 수정 노노 12. FCM 처리 - FCM을 통한 NOTIFICATION 완료 (1:1/그룹채팅) 13. 로그 아웃 처리 14. Setting 부분 처리중 15. 프로필 기본적인 처리완료(디버깅중) 16. 프로필 동기화 Glide처리(디버깅중) 17. 채팅대화목록 프로필 사진 처리 - 생명주기 고려

+++ app/src/main/java/com/example/nam/healthforyou/Setting.java
@@ -331,14 +331,19 @@ public class Setting extends AppCompatActivity implements View.OnClickListener{
                 if(extras != null)
                 {
                     Bitmap photo = extras.getParcelable("data");
-                    photo.compress(Bitmap.CompressFormat.JPEG, 100, stream);//이미지를 JPEG 형식으로 압축
+                    try{
+                        photo.compress(Bitmap.CompressFormat.JPEG, 90, stream);//이미지를 JPEG 형식으로 압축
+                    }catch(NullPointerException e)
+                    {
+                        Toast.makeText(mContext,"프로필 사진업로드 실패",Toast.LENGTH_SHORT).show();
+                    }
+
                     byte[] byteArray = stream.toByteArray();//스트림을 통해 bytearray로 만들고
                     Glide.with(this).load(byteArray).into(iv_myprofileImage);//Glide를 통해 이미지뷰에 올림
 
                     /////나의 프로필에 대한 이미지를 InternalStorage에 저장
-                    Bitmap bitmap = ((BitmapDrawable)iv_myprofileImage.getDrawable()).getBitmap();
                     new InternalImageManger(mContext).setFileName(myId).setDirectoryName("PFImage").save(photo);
-                    photo.compress(Bitmap.CompressFormat.JPEG, 100, stream);//이미지를 stream으로 옮김 - TODO 테스트 필요 두번 압축하기 때문에
+                    photo.compress(Bitmap.CompressFormat.JPEG, 90, stream);//이미지를 stream으로 옮김 - TODO 테스트 필요 두번 압축하기 때문에
                     byte[] byteArrayForupload = stream.toByteArray();//스트림을 통해 bytearray로 만들고
                     String base64Image = Base64.encodeToString(byteArrayForupload,Base64.DEFAULT);
 
@@ -373,11 +378,15 @@ public class Setting extends AppCompatActivity implements View.OnClickListener{
                 }
 
                 // 임시 파일 삭제
-                File f = new File(mImageCaptureUri.getPath());
-                if(f.exists())
+                if(mImageCaptureUri.getPath()!=null)
                 {
-                    f.delete();
+                    File f = new File(mImageCaptureUri.getPath());
+                    if(f.exists())
+                    {
+                        f.delete();
+                    }
                 }
+
 
                 break;
             }
